[RFC] ovsdb: Add aspirational documentation for clustering.
I'm calling this "aspirational documentation" because it documents code
that is incompletely implemented and does not work yet.  I'd appreciate
feedback on the plan.  I suggest reading ovsdb.7.xml in particular, which
explains the overall design, goal, and use of OVSDB, with a lot of detail
on how clustering is meant to work from a user perspective.

Signed-off-by: Ben Pfaff <blp@ovn.org>